Add optional higher order moments
These are part of the sums/sums_cov arrays returned
by gmix.get_weighted_moments

This is also available with GaussMom(fwhm, higher=True)

Currently the sums are calculated and are stored in a larger "sums" and
"sums_cov" arrays in the results.  No normalized moments are returned.

Names and indices for moments can be found in moments.MOMENTS_NAME_MAP