Fall back to casting to uint32
With numpy<2, bitwise_and does not operate on uint64 dtype objects.
Given that the range of fval is within the range of uint32, casting to
uint32 suffices. If the input value has more bits, they will be not used
anyway regardless of the casting as long as fval is limited to 2**31.